Move enum from types
