+++ src/TargetingData.ts
@@ -1,5 +1,4 @@
 import { NativeModules } from "react-native";
-import { UserGender, AdContentRating } from "./types/TargetingData";
 
 const { AppLovinMAX } = NativeModules;
 
@@ -22,6 +21,26 @@ interface NativeTargetingData {
 }
 
 const nativeModule: NativeTargetingData = AppLovinMAX;
+
+/**
+ * This enumeration represents content ratings for the ads shown to users.
+ */
+export enum AdContentRating {
+    None = 0,
+    AllAudiences = 1,
+    EveryoneOverTwelve = 2,
+    MatureAudiences = 3,
+}
+
+/**
+ * This enumeration represents gender.
+ */
+export enum UserGender {
+    Unknown = 'U',
+    Female = 'F',
+    Male = 'M',
+    Other = 'O',
+}
 
 export const TargetingData = {
