--- src/types/TargetingData.ts
@@ -1,19 +1,0 @@
-/**
- * This enumeration represents content ratings for the ads shown to users.
- */
-export enum AdContentRating {
-    None = 0,
-    AllAudiences = 1,
-    EveryoneOverTwelve = 2,
-    MatureAudiences = 3,
-}
-
-/**
- * This enumeration represents gender.
- */
-export enum UserGender {
-    Unknown = 'U',
-    Female = 'F',
-    Male = 'M',
-    Other = 'O',
-}
